requirements.txt
git+https://git.oschina.net/jjwt/Tbtools#egg=Tbtools
Flask==0.10.1
Flask-SQLAlchemy==2.1
celery[redis]
